+++ src/FileSystem/packageConfig.ts
@@ -1,0 +1,10 @@
+
+export interface ISosPackageConfig {
+	appletUid?: string;
+}
+
+export interface IPackageConfig {
+	name: string;
+	version?: string;
+	sos?: ISosPackageConfig;
+}